test: decouple RSS head link helpers
